[posix] fix options dependency (#5548)
A compile error can occur in the absence of LIBREADLINE. However the
error I got is confusing and doesn't indicate that because the options
dependencies are not good. I think the options' dependency and
relationship here should be:

- When DEAMON=0 (ot-cli, ot-ncp will be built and ot-cli depends on
  console), CLI_TRANSPORT=CONSOLE MUST be used. And otherwise
  (DEAMON=1), CLI_TRANSPORT=UART MUST be used. So when users build
  posix, they don't need to consider the option CLI_TRANSPORT. It's
  reasonable to add the evaluation for CLI_TRANSPORT in
  Makefile-posix.

- When CLI_TRANSPORT=CONSOLE, (HAVE_LIBEDIT || HAVE_LIBREADLINE) MUST
  be defined and DAEMON=0 MUST be satisfied. I think (HAVE_LIBEDIT ||
  HAVE_LIBREADLINE) is a result of CLI_TRANSPORT instead of the
  reason. The reason for CLI_TRANSPORT should be DEAMON.  Now I set
  CLI_TRANSPORT option based on DAEMON and make (HAVE_LIBEDIT ||
  HAVE_LIBREADLINE) as a check for CLI_TRANSPORT=CONSOLE.

If READLINE lib is not found, in this implementation, there will be an
error indicating that it's not found.